+++ development/http_client/dev_http_client.py
@@ -15,13 +15,9 @@ import pyjsonrpc
 rpc_client = pyjsonrpc.HttpClient("http://localhost:8080")
 
 # Example with *call*
-print rpc_client.call("add", 1, 2)
+print repr(rpc_client.call("add", 1, 2))
 
-# Example with direct *call*
-print rpc_client("add", 3, 4)
 
-# Example with *attribute* name
-print rpc_client.add(5, 6)
+# Example with *notify*
+print repr(rpc_client.notify("add", 1, 2))
 
-# Example with *notify* (no response)
-rpc_client.notify("add", 7, 8)