+++ examples/http_client_usage/multiple_example.py
@@ -14,14 +14,16 @@ import pyjsonrpc
 
 rpc_client = pyjsonrpc.HttpClient("http://localhost:8080")
 
-# Example with *call*
-print rpc_client.call("add", 1, 2)
+# Example with *multiple calls* in one request
+methods = [
+    pyjsonrpc.create_request_dict("add", 1, 2),
+    pyjsonrpc.create_request_dict("add", 3, 4)
+]
+print rpc_client.call(methods)
 
-# Example with direct *call*
-print rpc_client("add", 3, 4)
-
-# Example with *attribute* name
-print rpc_client.add(5, 6)
-
-# Example with *notify* (no response)
-rpc_client.notify("add", 7, 8)
+# Example with *multiple notifications* (no response) in one request
+notifications = [
+    pyjsonrpc.create_request_dict("add", 1, 2),
+    pyjsonrpc.create_request_dict("add", 3, 4)
+]
+rpc_client.notify(notifications)