+++ pyjsonrpc/http.py
@@ -13,6 +13,7 @@ import rpcrequest
 import rpcresponse
 import rpcerror
 import rpclib
+import Cookie
 from rpcjson import json
 
 import httplib
@@ -20,12 +21,38 @@ import gzip
 import StringIO
 
 
-def http_request(url, json_string, username=None, password=None, gzipped=False):
+
+def http_request(
+    url,
+    json_string,
+    username = None,
+    password = None,
+    timeout = None,
+    additional_headers = None,
+    content_type = None,
+    cookies = None,
+    gzipped=False
+):
     """
     Fetch data from webserver (POST request)
 
     :param json_string: JSON-String
+
     :param username: If *username* is given, BASE authentication will be used.
+
+    :param timeout: Specifies a timeout in seconds for blocking operations
+        like the connection attempt (if not specified, the global default
+        timeout setting will be used).
+        See: https://github.com/gerold-penz/python-jsonrpc/pull/6
+
+    :param additional_headers: Dictionary with additional headers
+        See: https://github.com/gerold-penz/python-jsonrpc/issues/5
+
+    :param content_type: Possibility to change the content-type header.
+
+    :param cookies: Possibility to add simple cookie-items as key-value pairs.
+        The key and the value of each cookie-item must be a bytestring.
+        Unicode is not allowed here.
     """
 
     request = urllib2.Request(url)
@@ -45,7 +72,7 @@ def http_request(url, json_string, username=None, password=None, gzipped=False):
 
     request.add_header("Content-Type", "application/json")
     if username:
-        base64string = base64.encodestring('%s:%s' % (username, password))[:-1]
+        base64string = base64.encodestring("%s:%s" % (username, password))[:-1]
         request.add_header("Authorization", "Basic %s" % base64string)
 
 
@@ -56,6 +83,18 @@ def http_request(url, json_string, username=None, password=None, gzipped=False):
 
     response = urllib2.urlopen(request)
 
+    # Cookies
+    if cookies:
+        cookie = Cookie.SimpleCookie(cookies)
+        request.add_header("Cookie", cookie.output(header = "", sep = ";"))
+
+    # Additional headers (overrides other headers)
+    if additional_headers:
+        for key, val in additional_headers.items():
+            request.add_header(key, val)
+
+    # Request
+    response = urllib2.urlopen(request, timeout = timeout)
 
     response_string = response.read()
 
@@ -69,6 +108,7 @@ def http_request(url, json_string, username=None, password=None, gzipped=False):
 
     response.close()
 
+    # Finished
     return response_string
 
 
@@ -105,6 +145,11 @@ class HttpClient(object):
         self.username = username
         self.password = password
         self.gzip_requests = gzipped
+        self.timeout = timeout
+        self.additional_headers = additional_headers
+        self.content_type = content_type
+        self.cookies = cookies
+
 
 
     def call(self, method, *args, **kwargs):
@@ -125,8 +170,8 @@ class HttpClient(object):
         if isinstance(method, basestring):
             request_json = rpcrequest.create_request_json(method, *args, **kwargs)
         else:
-            request_json = json.dumps(method)
             assert not args and not kwargs
+            request_json = json.dumps(method)
 
         # Call the HTTP-JSON-RPC server
         response_json = http_request(
@@ -135,20 +180,60 @@ class HttpClient(object):
             username = self.username,
             password = self.password,
             gzipped=self.gzip_requests
+            timeout = self.timeout,
+            additional_headers = self.additional_headers,
+            content_type = self.content_type,
+            cookies = self.cookies
+
         )
+        if not response_json:
+            return
 
         # Convert JSON-RPC-response to python-object
         response = rpcresponse.parse_response_json(response_json)
+        if isinstance(response, rpcresponse.Response):
+            if response.error:
+                # Raise error
+                raise rpcerror.jsonrpcerrors[response.error.code](
+                    message = response.error.message,
+                    data = response.error.data
+                )
+            else:
+                # Return result
+                return response.result
+        elif isinstance(response, list):
+            # Bei Listen wird keine Fehlerauswerung gemacht
+            return response
 
-        if response.error:
-            # Raise error
-            raise rpcerror.jsonrpcerrors[response.error.code](
-                message = response.error.message,
-                data = response.error.data
-            )
+
+    def notify(self, method, *args, **kwargs):
+        """
+        Sends a notification or multiple notifications to the server.
+
+        A notification is a special request which does not have a response.
+        """
+
+        methods = []
+
+        # Create JSON-RPC-request
+        if isinstance(method, basestring):
+            request_dict = rpcrequest.create_request_dict(method, *args, **kwargs)
+            request_dict["id"] = None
+            methods.append(request_dict)
         else:
-            # Return result
-            return response.result
+            assert not args and not kwargs
+            for request_dict in method:
+                request_dict["id"] = None
+                methods.append(request_dict)
+
+        # Redirect to call-method
+        self.call(methods)
+
+        # Fertig
+        return
+
+    # for compatibility with jsonrpclib
+    _notify = notify
 
 
     def __call__(self, method, *args, **kwargs):
@@ -182,7 +267,15 @@ class HttpRequestHandler(BaseHTTPServer.BaseHTTPRequestHandler, rpclib.JsonRpc):
     """
 
     protocol_version = "HTTP/1.1"
-    #server_version = "BaseHTTP/" + __version__
+    content_type = "application/json"
+
+
+    def set_content_type(self, content_type):
+        """
+        Set content-type to *content_type*
+        """
+
+        self.send_header("Content-Type", content_type)
 
 
     def set_content_type_json(self):
@@ -190,7 +283,7 @@ class HttpRequestHandler(BaseHTTPServer.BaseHTTPRequestHandler, rpclib.JsonRpc):
         Set content-type to "application/json"
         """
 
-        self.send_header("Content-Type", "application/json")
+        self.set_content_type("application/json")
 
 
     def set_no_cache(self):
@@ -216,7 +309,13 @@ class HttpRequestHandler(BaseHTTPServer.BaseHTTPRequestHandler, rpclib.JsonRpc):
         """
 
         # Parse URL query
-        query = urlparse.parse_qs(urllib.splitquery(self.path)[1])
+        path, query_str = urllib.splitquery(self.path)
+        if not query_str:
+            # Bad Request
+            return self.send_error(httplib.BAD_REQUEST)
+
+        # Parse querystring
+        query = urlparse.parse_qs(query_str)
 
         # jsonrpc
         jsonrpc = query.get("jsonrpc")
@@ -232,6 +331,9 @@ class HttpRequestHandler(BaseHTTPServer.BaseHTTPRequestHandler, rpclib.JsonRpc):
         method = query.get("method")
         if method:
             method = method[0]
+        else:
+            # Bad Request
+            return self.send_error(httplib.BAD_REQUEST)
 
         # params
         args = []
@@ -253,11 +355,11 @@ class HttpRequestHandler(BaseHTTPServer.BaseHTTPRequestHandler, rpclib.JsonRpc):
         request_json = json.dumps(request_dict)
 
         # Call
-        response_json = self.call(request_json)
+        response_json = self.call(request_json) or ""
 
         # Return result
         self.send_response(code = httplib.OK)
-        self.set_content_type_json()
+        self.set_content_type(self.content_type)
         self.set_no_cache()
         self.set_content_length(len(response_json))
         self.end_headers()
@@ -274,11 +376,11 @@ class HttpRequestHandler(BaseHTTPServer.BaseHTTPRequestHandler, rpclib.JsonRpc):
         request_json = self.rfile.read(content_length)
 
         # Call
-        response_json = self.call(request_json)
+        response_json = self.call(request_json) or ""
 
         # Return result
         self.send_response(code = httplib.OK)
-        self.set_content_type_json()
+        self.set_content_type(self.content_type)
         self.set_no_cache()
         self.set_content_length(len(response_json))
         self.end_headers()
@@ -319,7 +421,7 @@ def handle_cgi_request(methods = None):
                 args = []
                 kwargs = params
 
-        # Create JSON reqeust string
+        # Create JSON request string
         request_dict = rpcrequest.create_request_dict(method, *args, **kwargs)
         request_dict["jsonrpc"] = jsonrpc
         request_dict["id"] = id